module_3/requirements.txt
Flask==3.0.2
psycopg[binary]==3.1.18
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
huggingface_hub==0.14.1
transformers==4.31.0
llama-cpp-python==0.1.80


